Move Playbook::RedcarpetHelpere to Playbook::Markdown::Helper to be used within the library
