+++ playbook/lib/playbook/markdown.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require "playbook/markdown/template_handler"
+require "playbook/markdown/helper"
